+++ src/problem/MaxCoverage.kt
@@ -1,5 +1,6 @@
 package problem
 
+import data.Subsets
 import data.newName
 import discrete.Goal
 import discrete.Problem
@@ -11,13 +12,14 @@ fun newMaxCoverage(variant: String, n: Int): Problem? {
 	val name = newName(MaxCoverage, variant, n)
 	return when (variant) {
 		"basic" -> maxCoverage(name)
+		"weighted" -> weightedMaxCoverage(name)
 		else -> null
 	}
 }
 
-fun maxCoverage(name: String): Problem? {
+fun newMaxCoverageProblem(name: String): Pair<Problem?, Subsets?> {
 	val (p, cfg) = newSubsetsProblem(name)
-	if (p == null || cfg == null || cfg.limit == 0) return null
+	if (p == null || cfg == null || cfg.limit == 0) return Pair(null, null)
 
 	p.description += "\nLimit: ${cfg.limit}"
 	p.addUniversalConstraint(fun(solution: Solution): Boolean {
@@ -25,12 +27,34 @@ fun maxCoverage(name: String): Problem? {
 	})
 
 	p.goal = Goal.MAXIMIZE
+	return Pair(p, cfg)
+}
+
+fun maxCoverage(name: String): Problem? {
+	val (p, cfg) = newMaxCoverageProblem(name)
+	if (p == null || cfg == null) return null
+
 	p.objectiveFn = fun(solution: Solution): Score {
 		val covered = mutableSetOf<String>()
 		for(x in solution.asSubset()) {
 			covered.addAll(cfg.subsets[x])
 		}
 		return covered.size.toDouble()
+	}
+	return p
+}
+
+fun weightedMaxCoverage(name: String): Problem? {
+	val (p, cfg) = newMaxCoverageProblem(name)
+	if (p == null || cfg == null) return null
+	if (cfg.weight.size != cfg.universal.size) return null
+
+	p.objectiveFn = fun(solution: Solution): Score {
+		val covered = mutableSetOf<String>()
+		for(x in solution.asSubset()) {
+			covered.addAll(cfg.subsets[x])
+		}
+		return covered.sumOf { cfg.weight[it] ?: 0.0 }
 	}
 	return p
 }